unicorn restart in case pid not exists

+++ config/deploy.rb
@@ -57,8 +57,10 @@ namespace :unicorn do
   task :restart do
     on roles(:app) do 
       within("#{current_path}") do
-        pid = capture "cat #{current_path}/tmp/pids/unicorn.pid"
-        execute "kill -HUP #{pid}" if pid
+        if test("[ -f #{current_path}/tmp/pids/unicorn.pid ]")
+          pid = capture "cat #{current_path}/tmp/pids/unicorn.pid"
+          execute "kill -HUP #{pid}" if pid
+        end
         execute "uncorn_rails -c #{current_path}/config/unicorn.rb -D"
       end
     end
